Rename a utility class.
