+++ registration/src/test/java/org/spine3/samples/lobby/registration/seat/availability/SeatsAvailabilityRepositoryShould.java
@@ -39,8 +39,8 @@ import java.util.Map;
 
 import static org.junit.Assert.assertEquals;
 import static org.spine3.samples.lobby.common.util.IdFactory.*;
-import static org.spine3.samples.lobby.registration.util.MessageFactory.newSeatQuantity;
-import static org.spine3.samples.lobby.registration.util.MessageFactory.newSeatsAvailabilityId;
+import static org.spine3.samples.lobby.registration.util.Seats.newSeatQuantity;
+import static org.spine3.samples.lobby.registration.util.Seats.newSeatsAvailabilityId;
 
 /**
  * @author Alexander Litus
